Chat functionality is added
There are few things needed to update.
In the message screen. each chat is identified by the user ID, when we get into each chat screen , the chat screen is identify the each chat by the Uid, So right now its not showing any data its only shows a error null , due to users  in the chat screen are not defined the each uids, expect receive message and send message functionalists are added